// src/api/models/user-resource.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Role } from '../models/role';
export interface UserResource {
  created_at: (string | null);
  email: string;
  email_verified_at: (string | null);
  id: number;
  name: string;
  role: Role;
  updated_at: (string | null);
}